Dev: 分离dnspkg类
1、分离dnspkg类
2、重命名模块名
